cleanup: cleaned and seperated file structure
