+++ pyproject.toml
@@ -26,7 +26,7 @@ dev = [
 ]
 
 [project.scripts]
-sm = "sm.main:app"
+sm = "sm.app.main:app"
 
 [build-system]
 requires = ["uv_build>=0.7.21,<0.8.0"]
