+++ src/sm/tools/regex_generator/regex_sre_parse.py
